+++ app/src/main/java/eu/kanade/tachiyomi/data/track/anilist/AnilistModels.kt
@@ -9,7 +9,7 @@ import java.text.SimpleDateFormat
 import java.util.Locale
 
 data class ALManga(
-    val media_id: Int,
+    val media_id: Long,
     val title_user_pref: String,
     val image_url_lge: String,
     val description: String?,
