+++ app/src/main/java/eu/kanade/tachiyomi/data/track/model/TrackSearch.kt
@@ -10,7 +10,7 @@ class TrackSearch : Track {
 
     override var sync_id: Int = 0
 
-    override var media_id: Int = 0
+    override var media_id: Long = 0
 
     override var library_id: Long? = null
 
@@ -54,7 +54,7 @@ class TrackSearch : Track {
     override fun hashCode(): Int {
         var result = (manga_id xor manga_id.ushr(32)).toInt()
         result = 31 * result + sync_id
-        result = 31 * result + media_id
+        result = 31 * result + media_id.toInt()
         return result
     }
